Fix purchase item quantities to support decimals in API and related code
- Purchases API: return purchase_item_quantity and purchase_item_quantity_refunded
  as floats in both snapshot and fallback paths (enrichPurchaseItems)
- PosSessionsTable: use float for quantity in manual discount calculation
- ReceiptTemplateService: use float for quantity in VAT breakdown from items
- Add PurchasesApiTest asserting GET /api/purchases/{id} returns decimal quantity

+++ app/Filament/Resources/PosSessions/Tables/PosSessionsTable.php
@@ -1336,7 +1336,7 @@ class PosSessionsTable
 
                 // Count all discounts as manual for now
                 if ($discountAmount > 0) {
-                    $quantity = isset($item['quantity']) ? (int) $item['quantity'] : 1;
+                    $quantity = isset($item['quantity']) ? (float) $item['quantity'] : 1.0;
                     $manualDiscountCount++;
                     $manualDiscountAmount += $discountAmount * $quantity;
                 }
